Separate start and end caps, don't publicly export path_stroke and path_fill.
