+++ examples/gfx_advanced/src/main.rs
@@ -11,8 +11,8 @@ use lyon::path_builder::*;
 use lyon::math::*;
 use lyon::tessellation::geometry_builder::{VertexConstructor, VertexBuffers, BuffersBuilder};
 use lyon::tessellation::basic_shapes::*;
-use lyon::tessellation::path_fill::{FillTessellator, FillOptions};
-use lyon::tessellation::path_stroke::{StrokeTessellator, StrokeOptions};
+use lyon::tessellation::{FillTessellator, FillOptions};
+use lyon::tessellation::{StrokeTessellator, StrokeOptions};
 use lyon::tessellation;
 use lyon::path::Path;
 use lyon_renderer::buffer::{Id, BufferStore};
